updates BuyerDashboard
